fabio_bot/requirements.txt
# Fabio Bot - Order Flow Scalping (Bookmap add-on + optional standalone)
# Python 3.10+

# Core (used when running outside Bookmap, e.g. backtest/standalone)
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.28.0
websocket-client>=1.5.0

# Real market data for backtest (optional)
yfinance>=0.2.28

# Technical analysis (backtest / ATR)
pandas-ta>=0.3.14b0

# Optional: TA-Lib for ATR if available (pip install TA-Lib may need system lib)
# TA-Lib>=0.4.28

# API only (no dashboard)
fastapi>=0.104.0
uvicorn>=0.24.0

# ML signal filter (optional)
scikit-learn>=1.2.0
joblib>=1.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Bookmap API is provided by Bookmap when running inside Bookmap;
# no pip install for 'bookmap' - load add-on via Bookmap UI.
